refactor: useResponsivePageSize 훅 사용 및 코드 분리
