+++ src/pages/items-page/sections/BestProductions.jsx
@@ -1,8 +1,51 @@
-import { useState, useEffect, useCallback } from "react";
+import { useState, useEffect } from "react";
 import styled from "@emotion/styled";
 import { getItems } from "@apis/itemsApi";
 import ProductCard from "@components/ProductCard";
 import { breakpoints } from "@constants/breakpoints";
+import { useResponsivePageSize } from "@pages/items-page/hooks/useResponsivePageSize";
+
+const BestProductions = () => {
+  const [bestItemsData, setBestItemsData] = useState();
+  const pageSize = useResponsivePageSize({
+    desktop: 4,
+    tablet: 2,
+    mobile: 1,
+  });
+
+  useEffect(() => {
+    const fetchItems = async () => {
+      try {
+        const result = await getItems(1, pageSize, "favorite", "");
+        setBestItemsData(result.list);
+      } catch (error) {
+        console.error("Failed to fetch best items:", error);
+      }
+    };
+
+    fetchItems();
+  }, [pageSize]);
+
+  return (
+    <div>
+      <Title>베스트 상품</Title>
+      <ItemsContainer>
+        {bestItemsData?.map((item) => (
+          <ProductCard
+            key={item.id}
+            id={item.id}
+            src={item.images[0]}
+            title={item.name}
+            price={item.price}
+            like={item.favoriteCount}
+          ></ProductCard>
+        ))}
+      </ItemsContainer>
+    </div>
+  );
+};
+
+export default BestProductions;
 
 const Title = styled.h2`
   margin-bottom: 1.6rem;
@@ -25,57 +68,3 @@ export const ItemsContainer = styled.div`
     grid-template-columns: repeat(4, 1fr); /* PC: 5개 */
   }
 `;
-
-const BestProductions = () => {
-  const [bestItemsData, setBestItemsData] = useState();
-  const [count, setCount] = useState(1);
-
-  const updateCount = () => {
-    const width = window.innerWidth;
-    if (width >= parseInt(breakpoints.desktop)) {
-      setCount(4);
-    } else if (width >= parseInt(breakpoints.tablet)) {
-      setCount(2);
-    } else {
-      setCount(1);
-    }
-  };
-
-  const fetchItemsData = useCallback(async () => {
-    try {
-      const result = await getItems(1, count, "favorite", "");
-      setBestItemsData(result.list);
-    } catch (error) {
-      console.error("Failed to fetch donate data:", error);
-    }
-  }, [count]);
-
-  useEffect(() => {
-    updateCount();
-    window.addEventListener("resize", updateCount); // 브라우저 크기 변경 감지
-  }, []);
-
-  useEffect(() => {
-    if (count > 0) fetchItemsData();
-  }, [count, fetchItemsData]);
-
-  return (
-    <div>
-      <Title>베스트 상품</Title>
-      <ItemsContainer>
-        {bestItemsData?.map((item) => (
-          <ProductCard
-            key={item.id}
-            id={item.id}
-            src={item.images[0]}
-            title={item.name}
-            price={item.price}
-            like={item.favoriteCount}
-          ></ProductCard>
-        ))}
-      </ItemsContainer>
-    </div>
-  );
-};
-
-export default BestProductions;
